M2-4392: added code splitting for big components, increased lazy routers
