+++ src/modules/Auth/layouts/AuthLayout/index.ts
@@ -1,1 +1,2 @@
 export * from './AuthLayout';
+export { AuthLayout as default } from './AuthLayout';
