+++ src/modules/Builder/features/ActivityAbout/ActivityAbout.hooks.ts
@@ -4,7 +4,7 @@ import { useFormContext } from 'react-hook-form';
 import { getObjectFromList, getTextBetweenBrackets } from 'shared/utils';
 import { ItemFormValues } from 'modules/Builder/types';
 import { useCurrentActivity } from 'modules/Builder/hooks';
-import { checkIfItemHasRequiredOptions } from 'modules/Builder/features/ActivityItems/ItemConfiguration';
+import { checkIfItemHasRequiredOptions } from 'modules/Builder/features/ActivityItems/ItemConfiguration/ItemConfiguration.utils';
 
 const checkIfItemsHaveVariables = (items: ItemFormValues[]) => {
   const itemsObject = getObjectFromList(items, ({ name }) => name);
