+++ src/modules/Builder/features/ActivityItems/ItemConfiguration/index.ts
@@ -2,3 +2,4 @@ export * from './ItemConfiguration';
 export * from './ItemConfiguration.types';
 export * from './ItemConfiguration.const';
 export * from './ItemConfiguration.utils';
+export { ItemConfiguration as default } from './ItemConfiguration';
