+++ src/modules/Builder/features/ReportConfigSetting/index.ts
@@ -1,2 +1,3 @@
 export * from './ReportConfigSetting';
 export * from './ReportConfigSetting.types';
+export { ReportConfigSetting as default } from './ReportConfigSetting';
