+++ src/modules/Dashboard/features/Applet/index.ts
@@ -1,6 +1,3 @@
-export * from './AddUser';
-export * from './DashboardAppletSettings';
-export * from './Schedule';
 export * from './ShareApplet';
 export * from './TransferOwnership';
 export * from './Popups';
