+++ src/modules/Dashboard/features/Applets/index.ts
@@ -1,1 +1,2 @@
 export * from './Applets';
+export { Applets as default } from './Applets';
