+++ src/modules/Dashboard/features/Managers/index.ts
@@ -1,2 +1,3 @@
 export * from './Managers';
 export * from './Managers.types';
+export { Managers as default } from './Managers';
