+++ src/modules/Dashboard/features/RespondentData/RespondentDataReview/index.ts
@@ -1,2 +1,3 @@
 export * from './RespondentDataReview';
 export * from './RespondentDataReview.types';
+export { RespondentDataReview as default } from './RespondentDataReview';
