+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/index.ts
@@ -1,1 +1,2 @@
 export * from './RespondentDataSummary';
+export { RespondentDataSummary as default } from './RespondentDataSummary';
