+++ src/modules/Dashboard/features/Respondents/index.ts
@@ -1,2 +1,3 @@
 export * from './Respondents';
 export * from './Respondents.types';
+export { Respondents as default } from './Respondents';
