+++ src/modules/Dashboard/routes/routes.const.ts
@@ -1,10 +1,10 @@
 import { lazy } from 'react';
 
-import { Applets } from 'modules/Dashboard/features/Applets';
-import { Managers } from 'modules/Dashboard/features/Managers';
-import { Respondents } from 'modules/Dashboard/features/Respondents';
 import { page } from 'resources';
 
+const Applets = lazy(() => import('modules/Dashboard/features/Applets'));
+const Managers = lazy(() => import('modules/Dashboard/features/Managers'));
+const Respondents = lazy(() => import('modules/Dashboard/features/Respondents'));
 const AddUser = lazy(() => import('modules/Dashboard/features/Applet/AddUser'));
 const Schedule = lazy(() => import('modules/Dashboard/features/Applet/Schedule'));
 const AppletSettings = lazy(
