+++ src/shared/layouts/BaseLayout/components/TopBar/Notifications/Notification/Notification.tsx
@@ -6,7 +6,7 @@ import { generatePath, useNavigate } from 'react-router-dom';
 import { page } from 'resources';
 import { useEncryptionStorage } from 'shared/hooks';
 import { Svg } from 'shared/components/Svg';
-import { AppletPasswordPopup } from 'modules/Dashboard/features/Applet';
+import { AppletPasswordPopup } from 'modules/Dashboard/features/Applet/Popups';
 import { useAppDispatch } from 'redux/store';
 import { StyledLabelMedium, StyledLabelSmall, variables } from 'shared/styles';
 import { alerts } from 'shared/state';
