+++ src/shared/layouts/BaseLayout/index.tsx
@@ -1,1 +1,2 @@
 export * from './BaseLayout';
+export { BaseLayout as default } from './BaseLayout';
